refactor physics -> led, planck, math
