+++ src/lib.rs
@@ -6,15 +6,14 @@
 // To suppress these, `#![allow(clippy::approx_constant)]` is applied.
 #![allow(clippy::approx_constant)]
 ///#![allow(dead_code, unused_variables, unused_imports)]
-
 pub mod cam;
 pub mod colorant;
 mod error;
 pub mod geometry;
 pub mod illuminant;
 pub mod lab;
+pub mod math;
 pub mod observer;
-pub mod physics;
 pub mod prelude;
 pub mod rgb;
 pub mod spectrum;
