+++ src/spectrum/wavelength.rs
@@ -15,46 +15,8 @@
 //!
 
 use crate::spectrum::SPECTRUM_WAVELENGTH_RANGE;
-use num_traits::ToPrimitive;
 use core::f64;
-use std::{f64::consts::PI, sync::LazyLock};
-
-
-// calculated on first dereference, can not use floating point calculations in const (yet?)
-pub(crate) static FWHM: LazyLock<f64> = LazyLock::new(|| (8.0 * 2f64.ln()).sqrt());
-
-
-#[inline]
-pub fn sigma_from_fwhm(fwhm: f64) -> f64 {
-    fwhm / *FWHM
-}
-
-#[inline]
-pub fn fwhm_from_sigma(sigma: f64) -> f64 {
-    sigma * *FWHM
-}
-
-#[inline]
-pub fn gaussian_peak_one(x: f64, mu: f64, sigma: f64) -> f64 {
-    let exponent = -((x - mu).powi(2)) / (2.0 * sigma.powi(2));
-    exponent.exp()
-}
-
-#[test]
-fn gaussian_peak_one_test() {
-    use approx::assert_ulps_eq;
-    let sigma = 10E-9;
-    let mu = 500E-9;
-    let x = mu - sigma;
-    let v = gaussian_peak_one(x, mu, sigma);
-    assert_ulps_eq!(v, 0.60653065971, epsilon = 1E-10);
-}
-
-#[inline]
-pub fn gaussian_normalized(x: f64, mu: f64, sigma: f64) -> f64 {
-    let exponent = -((x - mu).powi(2)) / (2.0 * sigma.powi(2));
-    (1.0 / (sigma * (2.0 * PI).sqrt())) * exponent.exp()
-}
+use num_traits::ToPrimitive;
 
 #[inline(always)]
 pub fn wavelength<T: ToPrimitive>(i: T) -> f64 {
